requirements.txt
langchain-core
langchain-openai
python-dotenv
click
snowleopard